refactor: moved test files into same folder levels as sources
