+++ .meteor/packages
@@ -67,10 +67,7 @@ force-ssl@1.1.1
 # TESTING
 #========================================
 leaonline:testing
-# lmieulet:meteor-coverage@4.3.0
-# lmieulet:meteor-legacy-coverage@0.4.0
+lmieulet:meteor-coverage@5.0.0
 meteortesting:mocha@3.2.0
 # hwillson:stub-collections@2.0.0-rc.0
-
 # leaonline:status
-
