+++ .meteor/versions
@@ -54,6 +54,7 @@ leaonline:ratelimit-factory@2.0.0
 leaonline:testing@2.0.0
 leaonline:theme@2.0.0
 leaonline:utils@2.0.0
+lmieulet:meteor-coverage@5.0.0
 localstorage@1.2.1
 logging@1.3.5
 mdg:validated-method@1.3.0
