+++ package.json
@@ -17,10 +17,10 @@
     "@babel/runtime": "^7.24.8",
     "@fortawesome/fontawesome-free": "^5.15.4",
     "@popperjs/core": "^2.11.8",
-    "bcrypt": "^5.1.1",
-    "bootstrap": "^5.3.3",
+    "bcrypt": "^6.0.0",
+    "bootstrap": "^5.3.6",
     "jquery": "^3.7.1",
-    "meteor-node-stubs": "^1.2.9"
+    "meteor-node-stubs": "^1.2.19"
   },
   "meteor": {
     "mainModule": {
