+++ tests/api/server/contexts.tests.js
@@ -1,4 +1,4 @@
 /* eslint-env mocha */
 describe('contexts', function () {
-  import '../../../imports/api/admin/tests/Admin.tests'
+  import '../../../imports/api/admin/Admin.tests'
 })
